requirements.txt
mistralai
numpy
ollama>=0.5.4
openai
pillow
torch
pydantic
